ASPNET Core implementation.
