+++ .vs/restore.dg
@@ -1,1 +1,1 @@
-#:X:\Github-nlog-production\NLog.Web\NLog.Web.AspNetCore\NLog.Web.AspNetCore.xproj
+#:F:\Sreenath.Santhanam\GitHub\NLog.Web\NLog.Web.AspNetCore\NLog.Web.AspNetCore.xproj
