+++ NLog.Web.AspNetCore.Tests/LayoutRenderers/AspNetRequestReferrerRendererTests.cs
