+++ NLog.Web.AspNetCore/Enums/AspNetCookieLayoutOutPutFormat.cs
