+++ NLog.Web.Tests/NLog.Web.Tests.csproj
@@ -77,22 +77,22 @@
     <Otherwise />
   </Choose>
   <ItemGroup>
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\DefaultHttpContextAccessorTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\FakeHttpContextAccessor.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetApplicationValueLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetCookieLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetItemValueLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetRequestValueLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetSessionIDLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetSessionValueLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetUserAuthTypeLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\TestInvolvingAspNetHttpContext.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetUserIdentityLayoutRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetRequestReferrerRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetRequestUrlRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetRequestUserAgentTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetRequestHttpMethodRendererTests.cs" />
-    <Compile Include="..\NLog.Web.ASPNET5.Tests\LayoutRenderers\AspNetQueryStringLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\DefaultHttpContextAccessorTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\FakeHttpContextAccessor.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetApplicationValueLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetCookieLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetItemValueLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetRequestValueLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetSessionIDLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetSessionValueLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetUserAuthTypeLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\TestInvolvingAspNetHttpContext.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetUserIdentityLayoutRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetRequestReferrerRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetRequestUrlRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetRequestUserAgentTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetRequestHttpMethodRendererTests.cs" />
+    <Compile Include="..\NLog.Web.AspNetCore.Tests\LayoutRenderers\AspNetQueryStringLayoutRendererTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
